+++ app/pixel/pixel.go
@@ -8,14 +8,21 @@ import (
 	"sync"
 
 	"github.com/tarm/serial"
+	"io"
 )
 
-type Pixel struct{
+type Pixel interface {
+	SetState(pd PixelData)
+	GetState() (pd PixelData)
+}
+
+type SerialPixel struct{
 	sync.Mutex
 	PortName      string
 	PortSpeed     int
 	lastPixelData PixelData
-	serial        *serial.Port
+	Serial        io.Writer
+	Testing       bool
 }
 
 type PixelData struct {
@@ -25,7 +32,23 @@ type PixelData struct {
 	Brightness int
 }
 
-func (p *Pixel) Connect(){
+type StateCommand struct {
+	Pd          PixelData
+	PauseBefore int
+	PauseAfter  int
+}
+
+type NullWriter struct {
+	LastSerialMessage []byte
+}
+
+func (w NullWriter) Write (b []byte) (int, error){
+	w.LastSerialMessage = b
+	log.Printf("Write bytes: %s", b)
+	return 0, nil
+}
+
+func (p *SerialPixel) Connect(){
 	p.lastPixelData = PixelData{ -1, "", 0, 100 }
 
 	c := &serial.Config{Name: p.PortName, Baud: p.PortSpeed}
@@ -33,19 +56,18 @@ func (p *Pixel) Connect(){
 	if err != nil {
 		log.Fatalf("Could not open port %s, %s", c.Name, err)
 	}
-	p.serial = s
+	p.Serial = s
 
 	// port not opened before 1500 milliseconds pause
 	time.Sleep(1500 * time.Millisecond)
 }
 
-func (p Pixel) GetStatus() PixelData{
+func (p SerialPixel) GetState() PixelData{
 	return p.lastPixelData
 }
 
-func (p *Pixel) SetStatus(pd PixelData){
+func (p *SerialPixel) SetState(pd PixelData){
 	p.Lock()
-	animationDuration := 3000 // ms
 
 	switch pd.Blink {
 	case 1:
@@ -54,8 +76,29 @@ func (p *Pixel) SetStatus(pd PixelData){
 		pd.Value += 200
 	}
 
+	cmds := p.BuildSetStateCommands(pd)
+	if !p.Testing{
+		p.ExecCommands(cmds)
+	}
+
+	log.Printf("Pixel.SetState: %v\n", pd)
+
+	p.lastPixelData = pd
+
+	// if success value, turn off led
+	if pd.Value == 100{
+		p.lastPixelData = PixelData{ -1, "", 0, 100 }
+	}
+
+	p.Unlock()
+}
+
+func (p SerialPixel) BuildSetStateCommands (pd PixelData) ([]StateCommand){
+	cmds := []StateCommand{}
+	animationDuration := 3000 // ms
+
 	delta := pd.Value - p.lastPixelData.Value
-	stepTime := float64(animationDuration) / math.Abs(float64(delta))
+	stepTime := int(float64(animationDuration) / math.Abs(float64(delta)))
 
 	var step int
 	if delta > 0{
@@ -68,44 +111,66 @@ func (p *Pixel) SetStatus(pd PixelData){
 		// smooth switch color
 		if(pd.Blink == 0 && p.lastPixelData.Value > 0 && pd.Value > 0) {
 			for i := p.lastPixelData.Value; i != pd.Value; i += step {
-				p.sendSerial(PixelData{i, "", 0, pd.Brightness })
-				time.Sleep(time.Millisecond * time.Duration(stepTime))
+				cmds = append(cmds, StateCommand{
+					Pd: PixelData{i, "", 0, pd.Brightness },
+					PauseAfter: stepTime,
+				})
 			}
 		}
 	} else {
 		// sharp switch color
 		if(pd.Blink == 0 && p.lastPixelData.Value > 0 && pd.Value > 0) {
 			for i := 0; i < 3; i++ {
-				p.sendSerial(PixelData{pd.Value, "", 0, pd.Brightness })
-				time.Sleep(time.Millisecond * 250)
-				p.sendSerial(PixelData{p.lastPixelData.Value, "", 0, pd.Brightness })
-				time.Sleep(time.Millisecond * 250)
+				cmds = append(cmds, StateCommand{
+					Pd: PixelData{pd.Value, "", 0, pd.Brightness },
+					PauseAfter: 250,
+				})
+				cmds = append(cmds, StateCommand{
+					Pd: PixelData{p.lastPixelData.Value, "", 0, pd.Brightness },
+					PauseAfter: 250,
+				})
 			}
-			p.sendSerial(PixelData{pd.Value, "", 0, pd.Brightness })
+			cmds = append(cmds, StateCommand{
+				Pd: PixelData{pd.Value, "", 0, pd.Brightness },
+				PauseAfter: 250,
+			})
 		}
 	}
 
-	time.Sleep(100 * time.Duration(stepTime))
-	log.Printf("Pixel.setStatus: %v\n", pd)
-	p.sendSerial(pd)
-
-	p.lastPixelData = pd
+	cmds = append(cmds, StateCommand{
+		Pd: pd,
+		PauseAfter: 1000,
+	})
 
 	// if success value, turn off led
-	if pd.Value == 100{
-		time.Sleep(time.Millisecond * 5000)
-		p.sendSerial(PixelData{ -1, "", 0, 100 })
-		p.lastPixelData = PixelData{ -1, "", 0, 100 }
+	if pd.Value == 100 {
+		cmds = append(cmds, StateCommand{
+			Pd: PixelData{-1, "", 0, 100 },
+			PauseBefore: 5000,
+		})
 	}
 
-	time.Sleep(1000 * time.Millisecond)
-	p.Unlock()
+	return cmds
 }
 
-func (p Pixel) sendSerial (pd PixelData) (int, error){
+func (p SerialPixel) ExecCommands(cmds []StateCommand){
+	for _, c := range cmds{
+		if c.PauseBefore > 0{
+			time.Sleep(time.Millisecond * time.Duration(c.PauseBefore))
+		}
+
+		p.sendSerial(c.Pd)
+
+		if c.PauseAfter > 0{
+			time.Sleep(time.Millisecond * time.Duration(c.PauseAfter))
+		}
+	}
+}
+
+func (p SerialPixel) sendSerial (pd PixelData) (int, error){
 	command := fmt.Sprintf("%d|%s|%d\n",pd.Value, pd.Message, pd.Brightness)
 	//log.Println(command)
-	n, err := p.serial.Write([]byte(command))
+	n, err := p.Serial.Write([]byte(command))
 	if err != nil {
 		log.Fatalf("Could not write to port, %s", err)
 	}