+++ app/rest/server.go
@@ -13,7 +13,7 @@ import (
 
 type Server struct {
 	HostPort string
-	Pixel    pixel.Pixel
+	Pixel    pixel.SerialPixel
 }
 
 func (s Server) GetEngine() *gin.Engine {
@@ -25,13 +25,13 @@ func (s Server) GetEngine() *gin.Engine {
 	return router
 }
 
-func (s Server) Run() {
+func (s *Server) Run() {
 	gin.SetMode(gin.ReleaseMode)
 	router := s.GetEngine()
 
-	s.Pixel.SetStatus(pixel.PixelData{ 100, fmt.Sprintf("server started\\%s", s.HostPort), 1, 20 })
+	s.Pixel.SetState(pixel.PixelData{ 100, fmt.Sprintf("server started\\%s", s.HostPort), 1, 20 })
 	time.Sleep(2000 * time.Millisecond)
-	s.Pixel.SetStatus(pixel.PixelData{ -1, "", 0, 100 })
+	s.Pixel.SetState(pixel.PixelData{ -1, "", 0, 100 })
 
 	log.Fatal(router.Run(s.HostPort))
 }
@@ -48,10 +48,10 @@ func (s *Server) statusHandler(c *gin.Context) {
 		pd.Brightness = 100
 	}
 
-	go s.Pixel.SetStatus(pd)
+	go s.Pixel.SetState(pd)
 }
 
-func (s Server) kapacitorHandler(c *gin.Context) {
+func (s *Server) kapacitorHandler(c *gin.Context) {
 	ad := kapacitor.KapacitorAlertData{}
 	err := c.BindJSON(&ad)
 	if err != nil {
@@ -77,5 +77,5 @@ func (s Server) kapacitorHandler(c *gin.Context) {
 	data := ad.Data.Series[0]
 	pd.Message = fmt.Sprintf("%s\\%s: %v", data.Tags.Host, data.Name, data.Values[0][1]) // data.Values[1]
 
-	go s.Pixel.SetStatus(pd)
+	go s.Pixel.SetState(pd)
 }